fix(#201): address classifier round-4 gaps for builder split
- Merge serve_pipeline into run.rs; all run* entrypoints run on_shutdown hooks

- Remove new run_with_shutdown lib tests; restore pre-existing hot_reload unit test

- Move router/extract test bodies to tests/support/*_lib.rs via include!

- Delete src/router/tests/ directory and serve_pipeline.rs module

Verification: 226 lib tests pass (x2), all src/**/*.rs < 50KB, integration run coverage preserved
