+++ crates/rustapi-core/src/app/mod.rs
@@ -7,7 +7,6 @@ mod openapi;
 mod production;
 mod routing;
 mod run;
-mod serve_pipeline;
 mod types;
 
 #[cfg(test)]
