--- crates/rustapi-core/src/router/tests/mod.rs
@@ -1,3 +1,0 @@
-mod property_early;
-mod property_late;
-mod unit;
